Refactor out piece logic into a separate file.
